+++ controller/helpers.go
@@ -7,7 +7,6 @@ import (
 	enforcerconstants "go.aporeto.io/trireme-lib/controller/internal/enforcer/constants"
 	"go.aporeto.io/trireme-lib/controller/internal/processmon"
 	"go.aporeto.io/trireme-lib/controller/internal/supervisor/iptablesctrl"
-	"go.aporeto.io/trireme-lib/controller/pkg/claimsheader"
 	"go.aporeto.io/trireme-lib/controller/pkg/fqconfig"
 	"go.aporeto.io/trireme-lib/controller/pkg/packetprocessor"
 	"go.aporeto.io/trireme-lib/controller/pkg/remoteenforcer"
@@ -16,7 +15,7 @@ import (
 )
 
 // SetLogParameters sets up environment to be passed to the remote trireme instances.
-func SetLogParameters(logToConsole, logWithID bool, logLevel string, logFormat string, compressedTags claimsheader.CompressionType) {
+func SetLogParameters(logToConsole, logWithID bool, logLevel string, logFormat string, compressedTags constants.CompressionType) {
 
 	h := processmon.GetProcessManagerHdl()
 	if h == nil {
@@ -27,7 +26,7 @@ func SetLogParameters(logToConsole, logWithID bool, logLevel string, logFormat s
 }
 
 // GetLogParameters retrieves log parameters for Remote Enforcer.
-func GetLogParameters() (logToConsole bool, logID string, logLevel string, logFormat string, compressedTagsVersion claimsheader.CompressionType) {
+func GetLogParameters() (logToConsole bool, logID string, logLevel string, logFormat string, compressedTagsVersion constants.CompressionType) {
 
 	logLevel = os.Getenv(constants.EnvLogLevel)
 	if logLevel == "" {
@@ -44,12 +43,12 @@ func GetLogParameters() (logToConsole bool, logID string, logLevel string, logFo
 
 	logID = os.Getenv(constants.EnvLogID)
 
-	compressedTagsVersion = claimsheader.CompressionTypeNone
-	if console := os.Getenv(constants.EnvCompressedTags); console != string(claimsheader.CompressionTypeNone) {
-		if console == string(claimsheader.CompressionTypeV1) {
-			compressedTagsVersion = claimsheader.CompressionTypeV1
-		} else if console == string(claimsheader.CompressionTypeV2) {
-			compressedTagsVersion = claimsheader.CompressionTypeV2
+	compressedTagsVersion = constants.CompressionTypeNone
+	if console := os.Getenv(constants.EnvCompressedTags); console != string(constants.CompressionTypeNone) {
+		if console == string(constants.CompressionTypeV1) {
+			compressedTagsVersion = constants.CompressionTypeV1
+		} else if console == string(constants.CompressionTypeV2) {
+			compressedTagsVersion = constants.CompressionTypeV2
 		}
 	}
 	return
