+++ controller/internal/processmon/interfaces.go
@@ -1,8 +1,8 @@
 package processmon
 
 import (
+	"go.aporeto.io/trireme-lib/controller/constants"
 	"go.aporeto.io/trireme-lib/controller/internal/enforcer/utils/rpcwrapper"
-	"go.aporeto.io/trireme-lib/controller/pkg/claimsheader"
 	"go.aporeto.io/trireme-lib/policy"
 )
 
@@ -10,6 +10,6 @@ import (
 type ProcessManager interface {
 	KillProcess(contextID string)
 	LaunchProcess(contextID string, refPid int, refNsPath string, rpchdl rpcwrapper.RPCClient, arg string, statssecret string, procMountPoint string) (bool, error)
-	SetLogParameters(logToConsole, logWithID bool, logLevel string, logFormat string, compressedTags claimsheader.CompressionType)
+	SetLogParameters(logToConsole, logWithID bool, logLevel string, logFormat string, compressedTags constants.CompressionType)
 	SetRuntimeErrorChannel(e chan *policy.RuntimeError)
 }
