+++ controller/pkg/claimsheader/bytes.go
@@ -2,6 +2,8 @@ package claimsheader
 
 import (
 	"encoding/binary"
+
+	"go.aporeto.io/trireme-lib/controller/constants"
 )
 
 // HeaderBytess is the bytes returned
@@ -11,7 +13,7 @@ type HeaderBytes []byte
 func (c HeaderBytes) ToClaimsHeader() *ClaimsHeader {
 
 	claimsHeader := ClaimsHeader{}
-	claimsHeader.compressionType = CompressionTypeMask(c.parseClaimsHeaderAttr(CompressionTypeBitMask.ToUint8()))
+	claimsHeader.compressionType = constants.CompressionTypeMask(c.parseClaimsHeaderAttr(constants.CompressionTypeBitMask.ToUint8()))
 	claimsHeader.encrypt = Uint8ToBool(c.parseClaimsHeaderAttr(EncryptionEnabledMask))
 	claimsHeader.handshakeVersion = c.parseClaimsHeaderAttr(HandshakeVersion)
