+++ controller/pkg/claimsheader/bytes_test.go
@@ -4,13 +4,14 @@ import (
 	"testing"
 
 	. "github.com/smartystreets/goconvey/convey"
+	"go.aporeto.io/trireme-lib/controller/constants"
 )
 
 func TestHeaderBytes(t *testing.T) {
 
 	Convey("Given I create a new bytes", t, func() {
 		header := NewClaimsHeader(
-			OptionCompressionType(CompressionTypeV2Mask),
+			OptionCompressionType(constants.CompressionTypeV2Mask),
 			OptionEncrypt(true),
 			OptionHandshakeVersion(HandshakeVersion),
 		).ToBytes()
@@ -23,7 +24,7 @@ func TestHeaderBytes(t *testing.T) {
 			ch := header.ToClaimsHeader()
 
 			Convey("Then it should be equal", func() {
-				So(ch.compressionType, ShouldEqual, CompressionTypeV2Mask)
+				So(ch.compressionType, ShouldEqual, constants.CompressionTypeV2Mask)
 				So(ch.encrypt, ShouldEqual, true)
 				So(ch.handshakeVersion, ShouldEqual, HandshakeVersion)
 			})
