+++ controller/pkg/claimsheader/claimsheader_test.go
@@ -4,6 +4,7 @@ import (
 	"testing"
 
 	. "github.com/smartystreets/goconvey/convey"
+	"go.aporeto.io/trireme-lib/controller/constants"
 )
 
 func TestHeader(t *testing.T) {
@@ -11,7 +12,7 @@ func TestHeader(t *testing.T) {
 	Convey("Given I create a new claims header", t, func() {
 		header := NewClaimsHeader(
 			OptionEncrypt(true),
-			OptionCompressionType(CompressionTypeV2Mask),
+			OptionCompressionType(constants.CompressionTypeV2Mask),
 			OptionHandshakeVersion(HandshakeVersion),
 		).ToBytes()
 
@@ -23,7 +24,7 @@ func TestHeader(t *testing.T) {
 			ch := header.ToClaimsHeader()
 
 			Convey("Then it should be equal", func() {
-				So(ch.CompressionType(), ShouldEqual, CompressionTypeV2)
+				So(ch.CompressionType(), ShouldEqual, constants.CompressionTypeV2)
 				So(ch.Encrypt(), ShouldEqual, true)
 			})
 		})
@@ -32,7 +33,7 @@ func TestHeader(t *testing.T) {
 	Convey("Given I create a new claims header and encrypt false", t, func() {
 		header := NewClaimsHeader(
 			OptionEncrypt(false),
-			OptionCompressionType(CompressionTypeV1Mask),
+			OptionCompressionType(constants.CompressionTypeV1Mask),
 			OptionHandshakeVersion(HandshakeVersion),
 		).ToBytes()
 
@@ -44,7 +45,7 @@ func TestHeader(t *testing.T) {
 			ch := header.ToClaimsHeader()
 
 			Convey("Then it should be equal", func() {
-				So(ch.CompressionType(), ShouldEqual, CompressionTypeV1)
+				So(ch.CompressionType(), ShouldEqual, constants.CompressionTypeV1)
 				So(ch.Encrypt(), ShouldEqual, false)
 			})
 		})
@@ -56,7 +57,7 @@ func TestHeader(t *testing.T) {
 			ch := &ClaimsHeader{}
 
 			Convey("Then it should be equal", func() {
-				So(ch.CompressionType(), ShouldEqual, CompressionTypeNone)
+				So(ch.CompressionType(), ShouldEqual, constants.CompressionTypeNone)
 				So(ch.Encrypt(), ShouldEqual, false)
 			})
 		})
