+++ controller/pkg/secrets/compactpki.go
@@ -5,7 +5,7 @@ import (
 	"crypto/x509"
 	"errors"
 
-	"go.aporeto.io/trireme-lib/controller/pkg/claimsheader"
+	"go.aporeto.io/trireme-lib/controller/constants"
 	"go.aporeto.io/trireme-lib/controller/pkg/pkiverifier"
 	"go.aporeto.io/trireme-lib/utils/crypto"
 	"go.uber.org/zap"
@@ -17,7 +17,7 @@ type CompactPKI struct {
 	PublicKeyPEM  []byte
 	AuthorityPEM  []byte
 	TokenKeyPEMs  [][]byte
-	Compressed    claimsheader.CompressionType
+	Compressed    constants.CompressionType
 	privateKey    *ecdsa.PrivateKey
 	publicKey     *x509.Certificate
 	txKey         []byte
@@ -25,14 +25,14 @@ type CompactPKI struct {
 }
 
 // NewCompactPKI creates new secrets for PKI implementation based on compact encoding
-func NewCompactPKI(keyPEM []byte, certPEM []byte, caPEM []byte, txKey []byte, compress claimsheader.CompressionType) (*CompactPKI, error) {
+func NewCompactPKI(keyPEM []byte, certPEM []byte, caPEM []byte, txKey []byte, compress constants.CompressionType) (*CompactPKI, error) {
 
 	zap.L().Warn("DEPRECATED. secrets.NewCompactPKI is deprecated in favor of secrets.NewCompactPKIWithTokenCA")
 	return NewCompactPKIWithTokenCA(keyPEM, certPEM, caPEM, [][]byte{caPEM}, txKey, compress)
 }
 
 // NewCompactPKIWithTokenCA creates new secrets for PKI implementation based on compact encoding
-func NewCompactPKIWithTokenCA(keyPEM []byte, certPEM []byte, caPEM []byte, tokenKeyPEMs [][]byte, txKey []byte, compress claimsheader.CompressionType) (*CompactPKI, error) {
+func NewCompactPKIWithTokenCA(keyPEM []byte, certPEM []byte, caPEM []byte, tokenKeyPEMs [][]byte, txKey []byte, compress constants.CompressionType) (*CompactPKI, error) {
 
 	zap.L().Info("Initializing with Compact PKI")
 
@@ -165,7 +165,7 @@ type CompactPKIPublicSecrets struct {
 	CA          []byte
 	TokenCAs    [][]byte
 	Token       []byte
-	Compressed  claimsheader.CompressionType
+	Compressed  constants.CompressionType
 }
 
 // SecretsType returns the type of secrets.
